server/requirements.txt
Django==2.0.*
djangorestframework==3.7.*
django-cors-headers==2.2.*
social-auth-app-django==2.1.*
